# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "POVOD API",
    "version": "1.0.0",
    "description": "Контракт собирается из тех же Zod-схем, которые проверяют запросы и описывают ответы сервера, поэтому не может разойтись с реализацией (ARCH-002)."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальная разработка"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Резерв для клиентов без кук (VK Mini App, мобильные, тесты)."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "povod_session",
        "description": "Основной способ для браузера (SEC-001). Изменяющие запросы дополнительно требуют заголовок X-CSRF-Token со значением куки povod_csrf."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "EventPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/Events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Лента событий",
        "description": "Отдаётся страницами по курсору (BE-003). Для авторизованного пользователя события, совпадающие с его интересами, идут первыми.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Полнотекстовый поиск (BE-012)"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "author",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "startsFrom",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "name": "startsTo",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Непрозрачный курсор предыдущей страницы"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Страница ленты",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Создать событие",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 5000,
                    "default": ""
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "default": "Europe/Moscow"
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 300,
                    "default": ""
                  },
                  "category": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "image": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    },
                    "maxItems": 20
                  },
                  "coords": {
                    "type": "array",
                    "prefixItems": [
                      {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    ]
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "participantLimit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100000
                  }
                },
                "required": [
                  "title",
                  "startsAt"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Событие создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Нужна сессия",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/mine": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Мои события",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Созданные и посещаемые",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyEvents"
                }
              }
            }
          },
          "401": {
            "description": "Нужна сессия",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/favorites": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Избранные события",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Избранное текущего пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Нужна сессия",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Событие",
        "description": "Закрытое чужое событие отвечает 404, а не 403 — чтобы не подтверждать его существования. Параметр invite открывает ровно то событие, для которого выдан (BE-008).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "invite",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Событие",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Изменить событие",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 5000,
                    "default": ""
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "default": "Europe/Moscow"
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 300,
                    "default": ""
                  },
                  "category": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "image": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 50
                    },
                    "maxItems": 20
                  },
                  "coords": {
                    "type": "array",
                    "prefixItems": [
                      {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    ]
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "participantLimit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Событие обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или не пройдена проверка CSRF",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Удалить событие",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Удалено"
          },
          "403": {
            "description": "Недостаточно прав или не пройдена проверка CSRF",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{id}/join": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Записаться на событие",
        "description": "Повторный вызов идемпотентен. 409 — свободных мест не осталось (BE-007).",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "invite",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Запись подтверждена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или не пройдена проверка CSRF",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Мест не осталось",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Events/{id}/leave": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Отменить запись",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Запись отменена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Comments/event/{eventId}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Комментарии события",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "eventId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Комментарии",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}